kenzie_starter_app/serializers/creator_serializer.py
from rest_framework import serializers

from kenzie_starter_app.models import Creator


class CreatorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Creator
        fields = "__all__"
