requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.5.6
autopep8==1.5.7
backcall==0.2.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
dj-database-url==0.5.0
Django==3.1
django-filter==2.4.0
django-silk==4.1.0
djangorestframework==3.12.4
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==2.10
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.8.3
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
wrapt==1.12.1
